curd

+++ PT_KFC/QL_KFC/DTO/Combo_DTO.cs
@@ -10,14 +10,14 @@ namespace DTO
     {
         public string MaCombo { get; set; }
         public string TenCombo { get; set; }
-        public double GiaCombo { get; set; }
+        public int GiaCombo { get; set; }
         public int SoLuong { get; set; }
         public DateTime? NgayBatDau { get; set; }
         public DateTime? NgayKetThuc { get; set; }
 
         public Combo_DTO() { }
 
-        public Combo_DTO(string maCombo, string tenCombo, double giaCombo, int soLuong, DateTime? ngayBatDau, DateTime? ngayKetThuc)
+        public Combo_DTO(string maCombo, string tenCombo, int giaCombo, int soLuong, DateTime? ngayBatDau, DateTime? ngayKetThuc)
         {
             MaCombo = maCombo;
             TenCombo = tenCombo;
